Support  global install and running as CLI tool
