+++ package.json
@@ -2,7 +2,10 @@
   "name": "cht-upgrade-helper",
   "version": "4.0.0",
   "description": "Helper script for reviewing configuration when upgrading the CHT",
-  "main": "index.js",
+  "main": "./src/lib/main.js",
+  "bin": {
+    "cht-upgrade-helper": "src/bin/index.js"
+  },
   "scripts": {
     "eslint": "eslint 'src/**/*.js'"
   },
@@ -18,6 +21,7 @@
   "homepage": "https://github.com/jkuester/cht-upgrade-helper#readme",
   "dependencies": {
     "eslint": "^6.8.0",
+    "minimist": "^1.2.6",
     "xml2js": "^0.4.23"
   },
   "devDependencies": {
